+++ Tars/Tars/ViewController/SelectPlanetViewController.swift
@@ -9,18 +9,6 @@ import UIKit
 
 class SelectPlanetViewController: UIViewController {
     let contentsViewController = ContentsViewController()
-
-    let planetList: [Planet] = [
-        Planet(planetName: "태양", planetImage: UIImage(named: "Sun")),
-        Planet(planetName: "달", planetImage: UIImage(named: "Moon")),
-        Planet(planetName: "수성", planetImage: UIImage(named: "Mercury")),
-        Planet(planetName: "금성", planetImage: UIImage(named: "Venus")),
-        Planet(planetName: "화성", planetImage: UIImage(named: "Mars")),
-        Planet(planetName: "목성", planetImage: UIImage(named: "Jupiter")),
-        Planet(planetName: "토성", planetImage: UIImage(named: "Saturn")),
-        Planet(planetName: "천왕성", planetImage: UIImage(named: "Uranus")),
-        Planet(planetName: "해왕성", planetImage: UIImage(named: "Neptune"))
-    ]
     
     public let selectPlanetCollectionView: UICollectionView = {
         let layout = UICollectionViewFlowLayout()
@@ -41,9 +29,9 @@ class SelectPlanetViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        selectPlanetCollectionView.isUserInteractionEnabled = true
         selectPlanetCollectionView.delegate = self
         selectPlanetCollectionView.dataSource = self
+        
         view.backgroundColor = .white
         view.layer.masksToBounds = true
